Fixing PL after review
